Add files via upload
For DIRECTPyth Diagnostic Centers, there are more use-cases that are applicable
than those given here. However, to keep the process simple, we will be considering
only these four use-cases for our chatbot.

The healthcare industry is one of the most labor-intensive industries around the world. It
requires the presence of humans to take care of people at various stages of their illnesses.
One of friends was at the AI Conclave held by Amazon in 2017 in Bangalore and was
amazed to see how an acute problem of staff scarcity, which has been plaguing the
healthcare industry in the United Kingdom, has been aptly solved by creating artificial
tabletop bots that would take care of elderly patients needs.
The artificial tabletop bots remind elderly patients to take their pills, track their
prescriptions, and track and suggest wakeup routines. At the heart of Echo Alexa (as it is
known) is the machine learning developed by the Amazon team using its cloud
infrastructure Amazon Web Services (AWS). At the heart of Alexa is the Python machine
learning code that helps it to perform tasks and learn from them through a feedback
mechanism. The wonderful part of this service is that Echo Alexa is available to a common
Python developer to use and develop their own programs and products based on Amazon’s
infrastructure.